Added noodle customization pages.
