--- frontend/src/pages/About.js
@@ -1,7 +1,0 @@
-export default function Home(){
-    return(
-        <>
-        <h1>This is About</h1>
-        </>
-    )
-}
