--- frontend/src/pages/Home.js
@@ -1,12 +1,0 @@
-import './style.css'
-
-export default function Home(){
-    return(
-        <>
-        <div className='headdiv'>
-            <h1 className='headtext'>This is Home Page</h1>
-        </div>
-        <p>We Will Cook! Test</p>
-        </>
-    )
-}
