--- frontend/src/pages/style.css
@@ -1,11 +1,0 @@
-.headdiv{
-    background-color: chocolate;
-    text-align: center;
-}
-p{
-    text-align: center;
-    font-weight: 100;
-}
-.headtext{
-    font-family:'Franklin Gothic Medium';
-}
